add metainfo structs declaration
